requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
transformers>=4.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiohttp>=3.9.1
aiofiles>=23.2.1
